refactor: Moved filehandler to 'rest' package
closes #232
