+++ pkg/config/config.go
@@ -11,9 +11,9 @@ import (
 
 	protocolApi "github.com/trustbloc/sidetree-core-go/pkg/api/protocol"
 
-	"github.com/trustbloc/sidetree-fabric/pkg/filehandler"
 	"github.com/trustbloc/sidetree-fabric/pkg/rest/blockchainhandler"
 	"github.com/trustbloc/sidetree-fabric/pkg/rest/dcashandler"
+	"github.com/trustbloc/sidetree-fabric/pkg/rest/filehandler"
 )
 
 // DocumentType specifies the type of the Sidetree document
