+++ pkg/rest/filehandler/model.go
