+++ pkg/rest/filehandler/retrievehandler.go
