+++ pkg/rest/filehandler/retrievehandler_test.go
