+++ pkg/rest/filehandler/updateindexhandler.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package sidetreesvc
+package filehandler
 
 import (
 	"net/http"
@@ -13,29 +13,31 @@ import (
 	resthandler "github.com/trustbloc/sidetree-core-go/pkg/restapi/dochandler"
 )
 
-type fileIdxUpdateHandler struct {
+// UpdateIndex is a REST handler that creates/updates file index Sidetree documents
+type UpdateIndex struct {
 	*resthandler.UpdateHandler
 	path string
 }
 
-func newFileIdxUpdateHandler(path string, processor resthandler.Processor) *fileIdxUpdateHandler {
-	return &fileIdxUpdateHandler{
+// NewUpdateIndexHandler returns a new update index handler
+func NewUpdateIndexHandler(path string, processor resthandler.Processor) *UpdateIndex {
+	return &UpdateIndex{
 		path:          path,
 		UpdateHandler: resthandler.NewUpdateHandler(processor),
 	}
 }
 
 // Path returns the context path
-func (h *fileIdxUpdateHandler) Path() string {
+func (h *UpdateIndex) Path() string {
 	return h.path
 }
 
 // Method returns the HTTP method
-func (h *fileIdxUpdateHandler) Method() string {
+func (h *UpdateIndex) Method() string {
 	return http.MethodPost
 }
 
 // Handler returns the handler
-func (h *fileIdxUpdateHandler) Handler() common.HTTPRequestHandler {
+func (h *UpdateIndex) Handler() common.HTTPRequestHandler {
 	return h.Update
 }
