+++ pkg/rest/filehandler/updateindexhandler_test.go
@@ -4,7 +4,7 @@ Copyright SecureKey Technologies Inc. All Rights Reserved.
 SPDX-License-Identifier: Apache-2.0
 */
 
-package sidetreesvc
+package filehandler
 
 import (
 	"net/http"
@@ -13,12 +13,12 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestFileIdxResolveHandler(t *testing.T) {
+func TestUpdateIndex(t *testing.T) {
 	path := "/file"
 
-	h := newFileIdxResolveHandler(path, nil)
+	h := NewUpdateIndexHandler(path, nil)
 	require.NotNil(t, h)
-	require.Equal(t, "/file/{id}", h.Path())
+	require.Equal(t, "/file", h.Path())
 	require.NotNil(t, h.Handler())
-	require.Equal(t, http.MethodGet, h.Method())
+	require.Equal(t, http.MethodPost, h.Method())
 }
