+++ pkg/rest/filehandler/validator.go
