+++ pkg/rest/filehandler/validator_test.go
